chore: move helper class/methods/extension to package `helper`
